+++ include/libvideo2x/avutils.h
@@ -7,4 +7,7 @@ extern "C" {
 
 int64_t get_video_frame_count(AVFormatContext *ifmt_ctx, int in_vstream_idx);
 
+enum AVPixelFormat
+get_encoder_default_pix_fmt(const AVCodec *encoder, AVPixelFormat target_pix_fmt);
+
 #endif  // AVUTILS_H